Remove pytest-helpers package
This package does not play well with
type checkers, and it has not had
a release since 2021, so remove it.
