+++ integration_tests/conftest.py
@@ -8,7 +8,6 @@
 import os
 from collections.abc import Generator
 
-pytest_plugins = ["helpers_namespace"]
 import pytest
 from click.testing import CliRunner
 from datacube.cfg import ODCConfig
@@ -54,61 +53,6 @@ def ows_server(request, flask_app) -> WSGIServer:
 @pytest.fixture
 def runner() -> CliRunner:
     return CliRunner()
-
-
-@pytest.helpers.register
-def enclosed_bbox(
-    bbox: tuple[float, float, float, float], flip: bool = False
-) -> tuple[float, float, float, float]:
-    lon_min, lat_min, lon_max, lat_max = bbox
-    lon_range = lon_max - lon_min
-    lat_range = lat_max - lat_min
-
-    if flip:
-        return (
-            lat_min + 0.45 * lat_range,
-            lon_min + 0.45 * lon_range,
-            lat_max - 0.45 * lat_range,
-            lon_max - 0.45 * lon_range,
-        )
-    return (
-        lon_min + 0.45 * lon_range,
-        lat_min + 0.45 * lat_range,
-        lon_max - 0.45 * lon_range,
-        lat_max - 0.45 * lat_range,
-    )
-
-
-@pytest.helpers.register
-def disjoint_bbox(
-    bbox: tuple[float, float, float, float],
-) -> tuple[float, float, float, float]:
-    lon_min, lat_min, lon_max, lat_max = bbox
-    lon_range = lon_max - lon_min
-    lat_range = lat_max - lat_min
-
-    return (
-        lon_min - 0.4 * lon_range,
-        lat_min - 0.4 * lat_range,
-        lon_min - 0.2 * lon_range,
-        lat_min - 0.2 * lat_range,
-    )
-
-
-@pytest.helpers.register
-def representative_bbox(
-    bbox: tuple[float, float, float, float],
-) -> tuple[float, float, float, float]:
-    lon_min, lat_min, lon_max, lat_max = bbox
-    lon_range = lon_max - lon_min
-    lat_range = lat_max - lat_min
-
-    return (
-        lon_min + 0.40 * lon_range,
-        lat_min + 0.45 * lat_range,
-        lon_min + 0.41 * lon_range,
-        lat_min + 0.46 * lat_range,
-    )
 
 
 @pytest.fixture
